pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "EDMD_package"
version = "2.0.6"
authors = [
  { name="Márton Gadanecz", email="marton.gadanecz@gmail.com" },
  { name="Zsolt Fazekas", email="zsolt.fazekas.9@gmail.com" }
]
description="This module can be used to analyse the Phi and Psi dihedral (or torsion) angle distribution in a protein structural ensemble (e.g., from Chemical-Shift-Rosetta), define potential energy functions (PEFs), and replace the original dihedral energy terms in GROMACS for molecular dynamics (MD) simulations."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biopython~=1.81",
    "matplotlib~=3.7.1",
    "numpy~=1.26.4"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
license = "MIT"
license-files = ["LICENSE"]

[project.urls]
Homepage = "https://github.com/gadaneczm/EDMD_package"
